requirements.txt
langchain
python-dotenv
langchain_community
pypdf
wikipedia
langchain_text_splitters
langchain_openai
chromadb
faiss-cpu
langchain_chroma
streamlit
langchain-groq
langchain_core
uvicorn
fastapi
langserve
sse_starlette